Expose only one function and update tests
